fix: code refactor (#8)
* product eprel category mapping refactor

* code code-refactor

* code code-refactor
