--- src/main/java/it/gov/pagopa/register/dto/operation/KeyDTO.java
@@ -1,14 +1,0 @@
-package it.gov.pagopa.register.dto.operation;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class KeyDTO {
-
-  private String key;
-
-  private String message;
-
-}
