+++ src/main/java/it/gov/pagopa/register/event/consumer/ProductFileConsumerConfig.java
@@ -1,6 +1,6 @@
 package it.gov.pagopa.register.event.consumer;
 
-import it.gov.pagopa.register.service.operation.ProductFileService;
+import it.gov.pagopa.register.service.operation.ProductFileConsumerService;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.Message;
@@ -11,7 +11,7 @@ import java.util.function.Consumer;
 public class ProductFileConsumerConfig {
 
   @Bean
-  public Consumer<Message<String>> productFileConsumer(ProductFileService productFileService) {
-    return productFileService::execute;
+  public Consumer<Message<String>> productFileConsumer(ProductFileConsumerService productFileConsumerService) {
+    return productFileConsumerService::execute;
   }
 }
